remove warning on WIN32
git-svn-id: svn://svn.savannah.nongnu.org/osip/trunk/osip@171 a617929b-4c5a-41d8-b177-2eda7db2ca6c

+++ parser/urls.c
@@ -1013,7 +1013,7 @@ url_unescape (char *string)
 	  /* encoded part */
 	  if (sscanf (ptr + 1, "%02X", &hex))
 	    {
-	      in = hex;
+	      in = (int)hex;
 	      ptr += 2;
 	      alloc -= 2;
 	    }
